+++ src/statistics/ui/statistics-view.ts
@@ -60,6 +60,7 @@ export class StatisticsView extends ItemView {
 		return 'bar-chart-2';
 	}
 
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async onOpen(): Promise<void> {
 		this.service = new StatisticsService(this.plugin.app, this.plugin.settings);
 		this.stats = this.service.getAllStatistics();
@@ -68,6 +69,7 @@ export class StatisticsView extends ItemView {
 		this.registerEventHandlers();
 	}
 
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async onClose(): Promise<void> {
 		if (this.refreshTimeout) {
 			clearTimeout(this.refreshTimeout);
@@ -1576,6 +1578,7 @@ export class StatisticsView extends ItemView {
 		};
 	}
 
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async setState(state: Partial<StatisticsViewState>): Promise<void> {
 		if (state.expandedSections) {
 			this.expandedSections = new Set(state.expandedSections);